+++ plugins/search-backend-module-elasticsearch/src/engines/ElasticSearchSearchEngineIndexer.test.ts
@@ -176,6 +176,30 @@ describe('ElasticSearchSearchEngineIndexer', () => {
     expect(deleteSpy).toHaveBeenCalled();
   });
 
+  it('handles when no documents are received', async () => {
+    await TestPipeline.fromIndexer(indexer).withDocuments([]).execute();
+
+    // Older indices should have been queried for.
+    expect(catSpy).toHaveBeenCalled();
+
+    // A new index should have been created.
+    expect(createSpy).toHaveBeenNthCalledWith(
+      1,
+      expect.objectContaining({
+        path: expect.stringContaining('some-type-index__'),
+      }),
+    );
+
+    // No documents should have been sent
+    expect(bulkSpy).not.toHaveBeenCalled();
+
+    // Alias should not have been rotated.
+    expect(aliasesSpy).not.toHaveBeenCalled();
+
+    // Old index should not be cleaned up.
+    expect(deleteSpy).not.toHaveBeenCalled();
+  });
+
   it('handles bulk and batching during indexing', async () => {
     const documents = range(550).map(i => ({
       title: `Hello World ${i}`,