+++ plugins/sonarqube/src/components/SonarQubeContentPage/SonarQubeContentPage.test.tsx
@@ -22,8 +22,9 @@ import React from 'react';
 import {
   isSonarQubeAvailable,
   SONARQUBE_PROJECT_KEY_ANNOTATION,
-} from '../useProjectKey';
-import { SonarQubeApi, sonarQubeApiRef } from '../../api';
+  SonarQubeApi,
+  sonarQubeApiRef,
+} from '@backstage/plugin-sonarqube-react';
 
 const Providers = ({
   annotation,