+++ plugins/stack-overflow/src/search/StackOverflowSearchResultListItem/StackOverflowSearchResultListItem.tsx
@@ -26,17 +26,21 @@ import {
   Chip,
 } from '@material-ui/core';
 import { useAnalytics } from '@backstage/core-plugin-api';
+import { ResultHighlight } from '@backstage/plugin-search-common';
+import { HighlightedSearchResultText } from '@backstage/plugin-search-react';
 
 type StackOverflowSearchResultListItemProps = {
   result: any; // TODO(emmaindal): type to StackOverflowDocument.
   icon?: React.ReactNode;
   rank?: number;
+  highlight?: ResultHighlight;
 };
 
 export const StackOverflowSearchResultListItem = (
   props: StackOverflowSearchResultListItemProps,
 ) => {
   const { location, title, text, answers, tags } = props.result;
+  const { highlight } = props;
   const analytics = useAnalytics();
 
   const handleClick = () => {
@@ -55,10 +59,31 @@ export const StackOverflowSearchResultListItem = (
             primaryTypographyProps={{ variant: 'h6' }}
             primary={
               <Link to={location} noTrack onClick={handleClick}>
-                {_unescape(title)}
+                {highlight?.fields?.title ? (
+                  <HighlightedSearchResultText
+                    text={highlight.fields.title}
+                    preTag={highlight.preTag}
+                    postTag={highlight.postTag}
+                  />
+                ) : (
+                  _unescape(title)
+                )}
               </Link>
             }
-            secondary={`Author: ${text}`}
+            secondary={
+              highlight?.fields?.text ? (
+                <>
+                  Author:{' '}
+                  <HighlightedSearchResultText
+                    text={highlight.fields.text}
+                    preTag={highlight.preTag}
+                    postTag={highlight.postTag}
+                  />
+                </>
+              ) : (
+                `Author: ${text}`
+              )
+            }
           />
           <Chip label={`Answer(s): ${answers}`} size="small" />
           {tags &&