+++ packages/dev-utils/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/dev-utils
 
+## 1.0.9-next.4
+
+### Patch Changes
+
+- 2e701b3796: Internal refactor to use `react-router-dom` rather than `react-router`.
+- Updated dependencies
+  - @backstage/core-app-api@1.3.0-next.4
+  - @backstage/core-components@0.12.1-next.4
+  - @backstage/app-defaults@1.0.9-next.4
+  - @backstage/test-utils@1.2.3-next.4
+  - @backstage/plugin-catalog-react@1.2.2-next.4
+  - @backstage/catalog-model@1.1.4-next.1
+  - @backstage/core-plugin-api@1.2.0-next.2
+  - @backstage/integration-react@1.1.7-next.4
+  - @backstage/theme@0.2.16
+
 ## 1.0.9-next.3
 
 ### Patch Changes