+++ plugins/permission-node/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-permission-node
 
+## 0.7.2-next.3
+
+### Patch Changes
+
+- b05dcd5530: Move the `zod` dependency to a version that does not collide with other libraries
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.7.2-next.2
+  - @backstage/backend-common@0.17.0-next.3
+  - @backstage/config@1.0.5-next.1
+  - @backstage/errors@1.1.4-next.1
+  - @backstage/plugin-auth-node@0.2.8-next.3
+
 ## 0.7.2-next.2
 
 ### Patch Changes