+++ plugins/vault-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-vault-backend
 
+## 0.2.5-next.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-tasks@0.4.0-next.3
+  - @backstage/backend-common@0.17.0-next.3
+  - @backstage/backend-test-utils@0.1.31-next.4
+  - @backstage/config@1.0.5-next.1
+  - @backstage/errors@1.1.4-next.1
+
 ## 0.2.5-next.3
 
 ### Patch Changes